+++ database/migrations/2024_11_12_114212_create_order_items_table.php
@@ -15,9 +15,9 @@ return new class extends Migration
       $table->id();
       $table->foreignId('order_id')->constrained()->cascadeOnDelete();
       $table->foreignUuid('course_id')->constrained()->cascadeOnDelete();
-      $table->float('amount', 8, 2)->default(0);
-      $table->decimal('user_profit', 2)->default(0);
-      $table->decimal('platform_profit', 2)->default(0);
+      $table->decimal('amount', 8, 2)->default(0);
+      $table->decimal('user_profit', 8, 2)->default(0);
+      $table->decimal('platform_profit', 8, 2)->default(0);
       $table->boolean('withdraw')->default(false);
       $table->timestamps();
     });
